# crates/sito-api/src/router.rs
//! Axum router configuration binding handlers, middleware, and documentation.

use axum::extract::{Request, State};
use axum::http::{HeaderName, HeaderValue, Method, StatusCode, header};
use axum::middleware::Next;
use axum::response::{IntoResponse, Response};
use axum::routing::{delete, get, post, put};
use axum::{Json, Router};
use utoipa::OpenApi;
use utoipa_swagger_ui::SwaggerUi;

use crate::auth::client_ip::resolve_client_ip;
use crate::auth::session::extract_session_cookie;
use crate::auth::token::Role;
use crate::error::ProblemDetails;
use crate::handlers::{
    auth_handlers, cache, clients, config, filtering, ha, metrics, querylog, rewrites, stats,
    status, update, upstream,
};
use crate::openapi::ApiDoc;
use crate::state::ServerContext;

/// Header accepted for first-boot setup authorization.
pub const SETUP_TOKEN_HEADER: &str = "x-setup-token";

/// Middleware that enforces read-only access on replica slave nodes.
/// Mutating methods (POST, PUT, DELETE, PATCH) outside auth and resync return 409 Conflict with X-Dnsd-Master header.
pub async fn slave_read_only_middleware(
    axum::extract::State(ctx): axum::extract::State<ServerContext>,
    request: Request,
    next: Next,
) -> Response {
    let method = request.method();
    let is_mutating = matches!(
        *method,
        Method::POST | Method::PUT | Method::DELETE | Method::PATCH
    );

    let is_slave = ctx.config.load().server.role == "slave";

    if is_slave && is_mutating {
        let path = request.uri().path();
        let is_allowed = path.starts_with("/api/v1/auth")
            || path == "/api/v1/ha/resync"
            || path == "/ha/resync"
            || path == "/ui/login"
            || path == "/ui/logout"
            || path.starts_with("/auth");

        if !is_allowed {
            let master_url = ctx.resolve_master_url();
            let mut problem = ProblemDetails::conflict(
                "Modifications are disabled on HA replica slaves. Submit configuration changes directly to the master node.",
            );
            problem.error_type = "urn:sito:ha:read-only".to_string();
            problem.title = "Read-Only Replica".to_string();
            problem.instance = Some(path.to_string());

            let mut resp = (StatusCode::CONFLICT, Json(problem)).into_response();
            if !master_url.is_empty()
                && let Ok(val) = HeaderValue::from_str(&master_url)
            {
                resp.headers_mut()
                    .insert(HeaderName::from_static("x-dnsd-master"), val);
            }
            return resp;
        }
    }

    next.run(request).await
}

/// Extracts the setup token from the `X-Setup-Token` header or the
/// `setup_token` query parameter.
fn setup_token_from_request(request: &Request) -> Option<String> {
    if let Some(value) = request
        .headers()
        .get(SETUP_TOKEN_HEADER)
        .and_then(|v| v.to_str().ok())
        .map(str::trim)
        .filter(|v| !v.is_empty())
    {
        return Some(value.to_string());
    }
    let query = request.uri().query()?;
    for param in query.split('&') {
        let (key, value) = param.split_once('=').unwrap_or((param, ""));
        if key == "setup_token" && !value.is_empty() {
            return Some(value.to_string());
        }
    }
    None
}

/// Middleware that gates setup-only routes.
///
/// While a first-boot setup token is active, `/wizard`, `/ui/wizard/*` and
/// `/ui/upstreams/test` require either an authenticated session with
/// sufficient privileges or the one-time setup token. When `setup_pending` is
/// true the remaining bootstrapping behavior is preserved:
/// - Allows: `/static/*`, `/assets/*`, `/health`, `/status`.
/// - Returns 503 for `/api/v1/*`.
/// - Redirects all other UI routes to `/wizard`.
pub async fn setup_pending_middleware(
    State(ctx): State<ServerContext>,
    request: Request,
    next: Next,
) -> Response {
    let setup_pending = ctx.is_setup_pending();
    let token_required = ctx.auth_mgr.setup_token_required();

    if !setup_pending && !token_required {
        return next.run(request).await;
    }

    let path = request.uri().path();
    let is_wizard = path == "/wizard" || path.starts_with("/ui/wizard");
    let is_probe = path == "/ui/upstreams/test";
    if is_wizard || is_probe {
        let authenticated_role = request
            .headers()
            .get(header::COOKIE)
            .and_then(|v| v.to_str().ok())
            .and_then(extract_session_cookie)
            .and_then(|id| ctx.auth_mgr.validate_session(&id))
            .map(|session| session.role);
        let privileged = if is_probe {
            authenticated_role.is_some_and(|role| role >= Role::Operator)
        } else {
            authenticated_role == Some(Role::Admin)
        };

        if !privileged {
            // The wizard completion form carries the token in its body, which
            // would require buffering the request here; the handler performs
            // the same constant-time validation and rate limiting instead.
            if path == "/ui/wizard/complete" {
                return next.run(request).await;
            }
            if token_required {
                let config = ctx.config.load();
                let trusted_proxies = config.get_web_config().trusted_proxies.clone();
                let peer = request
                    .extensions()
                    .get::<axum::extract::ConnectInfo<std::net::SocketAddr>>()
                    .map(|ci| ci.0);
                let client_ip = resolve_client_ip(peer, request.headers(), &trusted_proxies);
                drop(config);
                let candidate = setup_token_from_request(&request);
                match ctx
                    .auth_mgr
                    .validate_setup_token(candidate.as_deref(), &client_ip)
                {
                    crate::auth::SetupTokenStatus::Valid => {}
                    crate::auth::SetupTokenStatus::RateLimited => {
                        return ProblemDetails::too_many_requests(
                            "Too many invalid setup token attempts. Try again later.",
                        )
                        .into_response();
                    }
                    _ => {
                        return ProblemDetails::forbidden(
                            "First-boot setup token required (see server logs).",
                        )
                        .into_response();
                    }
                }
            } else if !setup_pending && !ctx.auth_mgr.is_first_run() {
                // No token provisioned and not in a bootstrap state: let the
                // normal auth layer decide.
                return next.run(request).await;
            } else if !ctx.auth_mgr.is_first_run() {
                // Setup-pending with an already-configured admin account but
                // no setup token: require an authenticated admin session.
                return ProblemDetails::forbidden(
                    "Setup wizard is disabled. Admin session required.",
                )
                .into_response();
            }
        }
    }

    if setup_pending {
        if path.starts_with("/wizard")
            || path.starts_with("/ui/wizard")
            || path.starts_with("/static/")
            || path.starts_with("/assets/")
            || path == "/ui/upstreams/test"
            || path == "/health"
            || path == "/status"
        {
            return next.run(request).await;
        }

        if path.starts_with("/api/v1/") || path == "/api/v1" {
            return (StatusCode::SERVICE_UNAVAILABLE, "Setup not completed").into_response();
        }

        return (
            StatusCode::FOUND,
            [(axum::http::header::LOCATION, "/wizard")],
        )
            .into_response();
    }

    next.run(request).await
}

/// Unauthenticated liveness probe that also reports whether the first-run
/// setup wizard is still pending (used by the container healthcheck).
async fn health_handler(State(ctx): State<ServerContext>) -> impl IntoResponse {
    (
        StatusCode::OK,
        Json(serde_json::json!({
            "status": "ok",
            "setup_pending": ctx.is_setup_pending(),
        })),
    )
}

async fn not_found_handler() -> impl IntoResponse {
    (
        StatusCode::NOT_FOUND,
        Json(ProblemDetails::not_found(
            "The requested endpoint does not exist",
        )),
    )
}

/// Legacy alias for `GET|PUT /api/v1/upstream`; kept as a permanent redirect.
async fn redirect_upstream_config() -> Response {
    (
        StatusCode::PERMANENT_REDIRECT,
        [(axum::http::header::LOCATION, "/api/v1/upstream")],
    )
        .into_response()
}

/// Constructs the complete administrative HTTP router.
pub fn create_router(ctx: ServerContext) -> Router {
    let api_v1 = Router::new()
        // System status
        .route("/status", get(status::get_status))
        // Stats
        .route("/stats", get(stats::get_stats))
        .route("/stats/clients", get(stats::get_client_stats))
        .route("/stats/upstreams", get(stats::get_upstream_stats))
        // Query log
        .route(
            "/querylog",
            get(querylog::get_querylog).delete(querylog::delete_querylog),
        )
        .route("/querylog/clear", post(querylog::delete_querylog))
        .route("/querylog/stream", get(querylog::stream_querylog))
        // Filtering
        .route(
            "/filtering/lists",
            get(filtering::get_filter_lists).post(filtering::add_filter_list),
        )
        .route(
            "/filtering/lists/{id}",
            put(filtering::update_filter_list).delete(filtering::delete_filter_list),
        )
        .route("/filtering/refresh", post(filtering::refresh_filtering))
        .route(
            "/filtering/rules",
            get(filtering::get_filtering_rules).put(filtering::set_filtering_rules),
        )
        .route("/filtering/check", post(filtering::check_filtering))
        // Clients
        .route(
            "/clients",
            get(clients::get_clients).post(clients::create_client),
        )
        .route(
            "/clients/{name}",
            get(clients::get_client_by_name)
                .put(clients::update_client)
                .delete(clients::delete_client),
        )
        .route(
            "/clients/groups",
            get(clients::get_client_groups).post(clients::add_client_group),
        )
        .route(
            "/clients/groups/{name}",
            get(clients::get_client_group_by_name)
                .put(clients::update_client_group)
                .delete(clients::delete_client_group),
        )
        // Rewrites
        .route(
            "/rewrites",
            get(rewrites::get_rewrites).post(rewrites::add_rewrite),
        )
        .route(
            "/rewrites/{id}",
            put(rewrites::update_rewrite).delete(rewrites::delete_rewrite),
        )
        // Upstream: `/upstream` is canonical, `/upstream/config` redirects to it.
        .route(
            "/upstream",
            get(upstream::get_upstream_config).put(upstream::update_upstream_config),
        )
        .route(
            "/upstream/config",
            get(redirect_upstream_config).put(redirect_upstream_config),
        )
        .route("/upstream/test", post(upstream::test_upstream_servers))
        // Cache
        .route("/cache/flush", post(cache::flush_cache))
        .route("/cache/invalidate", post(cache::invalidate_cache))
        // High Availability (M8 stubs)
        .route("/ha/status", get(ha::get_ha_status))
        .route("/ha/slaves", get(ha::get_ha_slaves))
        .route("/ha/resync", post(ha::trigger_ha_resync))
        // Authentication
        .route("/auth/login", post(auth_handlers::login))
        .route("/auth/totp/verify", post(auth_handlers::verify_totp))
        .route("/auth/logout", post(auth_handlers::logout))
        .route("/auth/totp/setup", get(auth_handlers::get_totp_setup))
        .route("/auth/totp/enable", post(auth_handlers::enable_totp))
        .route("/auth/totp/disable", post(auth_handlers::disable_totp))
        .route(
            "/auth/tokens",
            get(auth_handlers::list_tokens).post(auth_handlers::create_token),
        )
        .route("/auth/tokens/{id}", delete(auth_handlers::delete_token))
        // Configuration
        .route(
            "/config",
            get(config::get_config).put(config::update_config),
        )
        .route("/config/reload", post(config::reload_config))
        .route("/config/backup", get(config::download_backup))
        .route("/config/restore", post(config::prepare_restore))
        .route("/config/restore/confirm", post(config::confirm_restore))
        // Software Update
        .route("/system/update/check", get(update::check_update))
        .route("/system/update/apply", post(update::apply_update))
        .layer(axum::middleware::from_fn_with_state(
            ctx.clone(),
            slave_read_only_middleware,
        ));

    // Note: Swagger UI at /api/docs and OpenAPI JSON are intentionally unauthenticated
    // to enable client generation and documentation discovery. Restrict via reverse proxy if needed.
    #[cfg(feature = "embed-ui")]
    let app = Router::new()
        .nest("/api/v1", api_v1)
        .route("/metrics", get(metrics::get_metrics))
        .route("/health", get(health_handler))
        .merge(SwaggerUi::new("/api/docs").url("/api/docs/openapi.json", ApiDoc::openapi()))
        .merge(
            crate::ui::ui_router().layer(axum::middleware::from_fn_with_state(
                ctx.clone(),
                slave_read_only_middleware,
            )),
        )
        .fallback(not_found_handler)
        .layer(axum::middleware::from_fn_with_state(
            ctx.clone(),
            setup_pending_middleware,
        ))
        .with_state(ctx.clone());

    #[cfg(not(feature = "embed-ui"))]
    let app = Router::new()
        .nest("/api/v1", api_v1)
        .route("/metrics", get(metrics::get_metrics))
        .route("/health", get(health_handler))
        .merge(SwaggerUi::new("/api/docs").url("/api/docs/openapi.json", ApiDoc::openapi()))
        .fallback(not_found_handler)
        .layer(axum::middleware::from_fn_with_state(
            ctx.clone(),
            setup_pending_middleware,
        ))
        .with_state(ctx.clone());

    app.layer(axum::middleware::from_fn_with_state(
        ctx.clone(),
        crate::security::csrf_protection_middleware,
    ))
    .layer(axum::middleware::from_fn_with_state(
        ctx,
        crate::security::security_headers_middleware,
    ))
}