# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sito DNS Administrative API",
    "description": "Administrative REST API and OpenAPI 3.0 specification for the sito DNS server",
    "license": {
      "name": "GPL-3.0-only"
    },
    "version": "1.6.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "crate::handlers::auth_handlers"
        ],
        "summary": "User login endpoint.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful or TOTP required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Locked out or rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/tokens": {
      "get": {
        "tags": [
          "crate::handlers::auth_handlers"
        ],
        "summary": "List all API tokens.",
        "operationId": "list_tokens",
        "responses": {
          "200": {
            "description": "List of API tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiTokenMeta"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "crate::handlers::auth_handlers"
        ],
        "summary": "Create a new API token.",
        "operationId": "create_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "API token created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/auth/tokens/{id}": {
      "delete": {
        "tags": [
          "crate::handlers::auth_handlers"
        ],
        "summary": "Revoke an API token by ID.",
        "operationId": "delete_token",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Token not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/auth/totp/disable": {
      "post": {
        "tags": [
          "crate::handlers::auth_handlers"
        ],
        "summary": "Disable TOTP 2FA. Requires the account password and, when TOTP is enabled,\na valid TOTP or backup code.",
        "operationId": "disable_totp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableTotpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "TOTP disabled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing second-factor code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid password or code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/auth/totp/enable": {
      "post": {
        "tags": [
          "crate::handlers::auth_handlers"
        ],
        "summary": "Enable TOTP 2FA by verifying setup code.",
        "operationId": "enable_totp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "TOTP enabled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/auth/totp/setup": {
      "get": {
        "tags": [
          "crate::handlers::auth_handlers"
        ],
        "summary": "Initiate TOTP 2FA setup.",
        "operationId": "get_totp_setup",
        "responses": {
          "200": {
            "description": "TOTP setup credentials generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotpSetupResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/auth/totp/verify": {
      "post": {
        "tags": [
          "crate::handlers::auth_handlers"
        ],
        "summary": "Second login phase: verify TOTP code.",
        "operationId": "verify_totp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "TOTP verified, session established",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cache/flush": {
      "post": {
        "tags": [
          "crate::handlers::cache"
        ],
        "summary": "Flush entire DNS response cache.",
        "operationId": "flush_cache",
        "responses": {
          "200": {
            "description": "Cache flushed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/cache/invalidate": {
      "post": {
        "tags": [
          "crate::handlers::cache"
        ],
        "summary": "Invalidate cache entries matching a domain.",
        "operationId": "invalidate_cache",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cache entries invalidated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/clients": {
      "get": {
        "tags": [
          "Clients"
        ],
        "operationId": "get_clients",
        "responses": {
          "200": {
            "description": "Clients list retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Clients"
        ],
        "operationId": "create_client",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Client created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/clients/groups": {
      "get": {
        "tags": [
          "Clients"
        ],
        "operationId": "get_client_groups",
        "responses": {
          "200": {
            "description": "Client groups retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientGroupDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Clients"
        ],
        "operationId": "add_client_group",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientGroupDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Client group added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientGroupDto"
                }
              }
            }
          },
          "400": {
            "description": "Group already exists"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/v1/filtering/check": {
      "post": {
        "tags": [
          "Filtering"
        ],
        "operationId": "check_filtering",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Filtering check result simulated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterCheckResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTokenRequest": {
        "type": "object",
        "description": "Request to create an API token.",
        "required": [
          "name",
          "scope"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          }
        }
      },
...
      "GenericMessageResponse": {
        "type": "object",
        "description": "Generic success message response.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "description": "Login request.",
        "required": [
          "user",
          "pass"
        ],
        "properties": {
          "pass": {
            "type": "string"
          },
          "user": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "description": "Login response.",
        "properties": {
          "partial_token": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "type": [
              "string",
              "null"
            ]
          },
          "session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "totp_required": {
            "type": "boolean"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}